Crate gabelstaplerwm

gabelstaplerwm - a semidynamic tiling window manager

It's what the heading says it is. The documentation found here is a very dense description of what the sources do. It is intended to be read along with the sources, as the configuration process involves you editing them. See the documentation for the config module for more information on configuration.

Modules

wm

Macros

bind
Bind a key combination to a callback closure.

edit_layout
Edit the current layout via a LayoutCommand.

focus
Focus a client using a closure.

move_to_tag
Move a client to a tag.

push_tagset
View a tagset by pushing it by index on the history stack.

set
Construct a Set of... things, like you would use vec!.

set_from_slice
Construct a Set of... things from a slice, like you would use slice.to_vec().

swap
Swap a client using a closure.

toggle_show_tag
Toggle a tag on the current tagset.

toggle_tag
Toggle a tag on a client.