Macro gabelstaplerwm::toggle_tag

macro_rules! toggle_tag {
    ($tag:expr;; $print:expr) => { ... };
    ($tag:expr $(; $print:expr)*) => { ... };
}

Toggle a tag on a client.

Returns a closure for use with bind!.

Usage

The toggle_tag! macro expects a tag to be toggled on the currently focused client.

As always, the last parameter(s) specify objects to be printed after completion of the action.