Struct gabelstaplerwm::wm::client::Client

pub struct Client {
    pub window: Window,
    // some fields omitted
}

A client wrapping a window.

A client is a container object that holds the information associated with, a window, but doesn't directly influence the workings of the window manager. That is, the window's properties are used to alter associated structures, which in turn influence the behaviour of the window manager. This is a common pattern in gabelstaplerwm: Most code that the user writes isn't calling any X functions to actually perform the actions it symbolizes. Instead, it modifies carefully crafted structures that act as an interpreting layer for the window manager.

Fields
window: Window

the window (a direct child of root)

Methods

impl Client

fn new(window: Window, tags: BTreeSet<Tag>, props: ClientProps) -> Client

Setup a new client for a specific window, on a set of tags and with given properties.

fn set_tags(&mut self, tags: &[Tag])

Move a window to a new set of tags.

Assumes the slice denoted by tags doesn't contain duplicate elements.

fn toggle_tag(&mut self, tag: Tag) -> Option<bool>

Add or remove a tag from a window.

If client would be visible on no tags at all, the operation is not performed.

fn match_tags(&self, tags: &BTreeSet<Tag>) -> bool

Check whether a client is visible on a set of tags.