Struct gabelstaplerwm::wm::client::ScreenSet

pub struct ScreenSet { /* fields omitted */ }

An ordered set of known screens.

A screen is a rectangular area on the X server screen's root window, that is used to show a distinct set of tags associated with a TagStack. There is an active screen at all times.

Methods

impl ScreenSet

fn new(screens: HashMap<Crtc, Screen>, crtcs: Vec<Crtc>) -> Option<ScreenSet>

Setup a new screen set.

fn current_mut(&mut self) -> &mut Screen

Get a mutable reference to current screen's geometry and tag stack.

fn current(&self) -> &Screen

Get an immutable reference to current screen's geometry and tag stack.

fn screen(&self) -> &TilingArea

Get an immutable reference to current screen's geometry.

fn tag_stack_mut(&mut self) -> &mut TagStack

Get a mutable reference to the current screen's tag stack.

fn tag_stack(&self) -> &TagStack

Get an immutable reference to the current screen's tag stack.

fn rotate(&mut self)

Swap horizontal and vertical axes of all screens.

fn select_screen(&mut self, new: Crtc) -> bool

Select a screen by index.

fn remove(&mut self, crtc: Crtc)

Remove a CRTC from our list of screens.

fn run_matching(&mut self, matching: &ScreenMatching)

Apply a screen matching to all screens (that is, CRTCs) that we know of.

fn update(&mut self, change: &CrtcChange)

Update a screen associated with a CRTC or create one if none is present.