Struct gabelstaplerwm::wm::client::TagSet

pub struct TagSet {
    pub layout: Box<Layout>,
    // some fields omitted
}

A set of tags with an associated layout.

Used to determine the set of windows to be shown, as well as their geometries. All clients that match any of the tags in a tagset are shown to the user when that tagset is displayed by the window manager. In that sense, tagsets are views into the space of open clients, with additional parameters attached. Note that layouts are dynamically determined and specified by a trait object, allowing for easy extending of the defaults.

Fields
layout: Box<Layout>

the layout used to display clients on the tagset

Methods

impl TagSet

fn new<L: Layout + 'static>(tags: BTreeSet<Tag>, layout: L) -> TagSet

Initialize a new tag set with a layout and a set of tags.

fn toggle_tag(&mut self, tag: Tag) -> bool

Toggle a tag on the tagset and return which changes have been made.

fn replace_tag(&mut self, old: Tag, new: Tag)

Replace a tag in the tagset, if it is present.

fn get_tags(&self) -> &BTreeSet<Tag>

Get the tag set from the tag set. Funny.