Struct gabelstaplerwm::wm::client::TagStack

pub struct TagStack { /* fields omitted */ }

An organized set of known tagsets.

Allows for simple addressing of tagstes (and their layouts) Tagsets are added and removed using API calls and are adressed using 8-bit unsigned integers. Thus, 256 different tagsets can be managed at any point in time. A small history of capped size is kept, determining the tagset currently displayed by the window manager.

Methods

impl TagStack

fn new() -> TagStack

Setup an empty tag stack.

fn from_presets(vec: Vec<TagSet>, viewed: u8) -> TagStack

Setup a tag stack from a vector of tag sets and the index of the initially viewed tagset in the vector.

fn current_index(&self) -> Option<&u8>

Get the current tag set's index

Returns None if the history stack is empty

fn current(&self) -> Option<&TagSet>

Get the current tag set by reference.

Returns None if the history stack is empty

fn current_mut(&mut self) -> Option<&mut TagSet>

Get the current tag set by mutable reference.

Returns None if the history stack is empty

fn push(&mut self, new_index: u8)

Set the currently viewed tagset by index.

fn add(&mut self, index: u8, value: TagSet) -> bool

Add a new tagset to the set.

fn remove(&mut self, index: u8) -> bool

Remove a tagset from the set.

fn view_prev(&mut self) -> bool

Switch to previously shown tagset, using the history stack.