Enum gabelstaplerwm::wm::err::WmError

pub enum WmError {
    CouldNotEstablishHandlers,
    CouldNotConnect(ConnError),
    CouldNotAcquireScreen,
    CouldNotRegisterAtom(String),
    OtherWmRunning,
    RandRVersionMismatch,
    RandRSetupFailed,
    ConnectionInterrupted,
    IOError,
}

An error encountered by the WM.

Variants
CouldNotEstablishHandlers

Signal handlers handling SIGCHLD's can't be established.

CouldNotConnect(ConnError)

Could not connect to X server.

CouldNotAcquireScreen

Could not acquire a screen reference from the X server.

CouldNotRegisterAtom(String)

An atom used by the window manager wasn't accepted by the X server.

OtherWmRunning

Another window manager is running, so we can't get the necessary event mask registered with the X server.

RandRVersionMismatch

The RandR version we need is not supported.

RandRSetupFailed

The call to RandR initializing event receival failed.

ConnectionInterrupted

The connection to the X server has been interrupted.

IOError

Input/Output with the X server has issues.

Methods

impl WmError

fn handle(self) -> !

"Handle" an error, ie. print error message and exit.