Enum gabelstaplerwm::wm::layout::LayoutMessage

pub enum LayoutMessage {
    MasterFactorAbs(u8),
    MasterFactorRel(i8),
    FixedAbs(bool),
    FixedRel,
    XOffAbs(u32),
    XOffRel(i32),
    YOffAbs(u32),
    YOffRel(i32),
    ColumnAbs(u8),
    ColumnRel(i8),
}

A message type being sent to layout objects.

Introduced to allow for type- and implementation-independent layout editing from keybindings and other code. Layout implementations can choose to react to any subset of the message variants below, or none at all.

Variants
MasterFactorAbs(u8)

Set absolute value of the master factor.

MasterFactorRel(i8)

Add an offset to the master factor.

FixedAbs(bool)

Set fixed attribute of layout.

FixedRel

Toggle fixed attrbute of layout.

XOffAbs(u32)

Set absolute value of the x offset.

XOffRel(i32)

Add an offset to the x offset.

YOffAbs(u32)

Set absolute value of the y offset.

YOffRel(i32)

Add an offset to the y offset.

ColumnAbs(u8)

Set absolute value of the column amount.

ColumnRel(i8)

Add an offset to the column amount.