Struct gabelstaplerwm::wm::layout::stack::HStack

pub struct HStack {
    pub master_factor: u8,
    pub inverted: bool,
    pub fixed: bool,
}

Horizontal stack layout.

+-------+
|   A   | A: master window
+-+-+-+-+
| |B| | | B: stack, hidden if fixed=false and num_windows <= 1
+-+-+-+-+

New windows are added as slaves to the stack.

Fields
master_factor: u8

percentage of screen height taken by the master window area, saturating semantics

inverted: bool

place the stack on top?

fixed: bool

keep the height(s) of the areas even if they are empty?