Enum gabelstaplerwm::wm::window_system::WmCommand

pub enum WmCommand {
    Redraw,
    Focus,
    Kill(Window),
    ModeSwitch(Mode),
    Quit,
    NoCommand,
}

Enumeration type of commands executed by the window manager.

Being returned from a callback closure which modified internal structures, gets interpreted to take necessary actions.

Variants
Redraw

redraw everything

Focus

reset focus

Kill(Window)

kill the client associated with the window

ModeSwitch(Mode)

switch keyboard mode

Quit

quit window manager

NoCommand

don't do anything, no action is needed