Struct gabelstaplerwm::wm::window_system::Wm

pub struct Wm<'a> { /* fields omitted */ }

A window manager master-structure.

This is the central instance coordinating the communication with the X server, as well as containing structures to manage tags and clients. It also contains callback mechanisms upon key press and client creation.

Methods

impl<'a> Wm<'a>

fn new(
con: &'a Connection,
screen_num: i32,
config: WmConfig
) -> Result<Wm<'a>, WmError>

Wrap a connection to initialize a window manager.

fn init_randr(&mut self) -> Result<(), WmError>

Initialize the RandR extension for multimonitor support.

fn init_clients(&mut self)

Add all present clients to the datastructures on startup.

fn register(&mut self) -> Result<(), WmError>

Register window manager.

Issues substructure redirects for the root window and registers for all events we are interested in.

fn setup_bindings(&mut self, keys: Vec<(KeyPress, KeyCallback)>)

Set up keybindings and necessary keygrabs.

fn setup_matching(&mut self, matching: Matching)

Set up client matching.

fn setup_screen_matching(&mut self, matching: ScreenMatching)

Set up screen matching.

fn setup_urgency_callback(&mut self, callback: UrgencyCallback)

Set up urgency callback.

fn run(&mut self) -> Result<(), WmError>

Wait for events, handle them. Repeat.