Struct gabelstaplerwm::wm::client::TagSet

pub struct TagSet {
    pub tags: BTreeSet<Tag>,
    pub layout: Box<Layout>,
}

A set of tags with an associated layout.

Used to determine the set of windows to be shown, as well as their geometries. All clients that match any of the tags in a tagset are shown to the user when that tagset is displayed by the window manager. In that sense, tagsets are views into the space of open clients, with additional parameters attached. Note that layouts are dynamically determined and specified by a trait object, allowing for easy extending of the defaults.

Fields
tags: BTreeSet<Tag>

tags belonging to tagset

layout: Box<Layout>

the layout used to display clients on the tagset

Methods

impl TagSet

fn new<L: Layout + 'static>(tags: BTreeSet<Tag>, layout: L) -> TagSet

Initialize a new tag set with a layout and a set of tags.

fn toggle_tag(&mut self, tag: Tag) -> bool

Toggle a tag on the tagset and return whether changes have been made.

fn set_layout<L: Layout + 'static>(&mut self, layout: L)

Set a layout on the tagset.